Pass through mouse down event when starting selection (#123)
When a renderable is `selectable`, it will start the selection, but pass
through the `mousedown` event, so when releasing the button immediately,
no selection happens and `mousedown` can be handled by the renderable.
To avoid actually selecting the text in a text renderable one still has
to set `selectable=false`, for a button component for example. So you
can have both.

Should solve #112